+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## Version 1.3.6 - Community 1.0.0
 
     * [SER-3363] - Implement label translation in Client related UIs
+    * [SER-3259] - External Service: Add new page for payment provider configuration
 
 ## Version 1.3.5 - Community 1.0.0